+++ src/stats/combos.ts
@@ -1,5 +1,5 @@
 import _ from "lodash";
-import { PostFrameUpdateType } from "../utils/slpReader";
+import { PostFrameUpdateType } from "../types";
 import { FrameEntryType, FramesType, MoveLandedType, ComboType, PlayerIndexedType } from "./common";
 import { isDamaged, isGrabbed, calcDamageTaken, isTeching, didLoseStock, Timers, isDown, isDead } from "./common";
 import { StatComputer } from "./stats";
