--- src/utils/index.ts
@@ -1,11 +1,0 @@
-export * from "./slpFile";
-export {
-  PostFrameUpdateType,
-  GameStartType,
-  GameEndType,
-  Command,
-  PreFrameUpdateType,
-  PlayerType,
-  parseMessage,
-} from "./slpReader";
-export { SlpParser } from "./slpParser";
